requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
dj-database-url==0.5.0
Django==1.11.5
django-bootstrap4==0.0.6
gunicorn==19.8.1
idna==2.6
Pillow==5.1.0
psycopg2==2.7.4
pyasn1==0.4.2
pycparser==2.18
python-decouple==3.1
pytz==2018.4
six==1.11.0
urllib3==1.22
whitenoise==3.3.1
